requirements.txt
streamlit
PyPDF2
pandas
numpy
matplotlib
seaborn
Pillow
pytesseract 